[Optimization-439][client] Optimize CDCSource sync doris
