+++ dlink-client/dlink-client-1.13/src/main/java/com/dlink/cdc/jdbc/JdbcSinkBuilder.java
@@ -1,5 +1,26 @@
 package com.dlink.cdc.jdbc;
 
+import org.apache.flink.api.dag.Transformation;
+import org.apache.flink.connector.jdbc.dialect.ClickHouseDialect;
+import org.apache.flink.connector.jdbc.dialect.MySQLDialect;
+import org.apache.flink.connector.jdbc.dialect.OracleDialect;
+import org.apache.flink.connector.jdbc.dialect.PostgresDialect;
+import org.apache.flink.connector.jdbc.dialect.SQLServerDialect;
+import org.apache.flink.connector.jdbc.internal.options.JdbcOptions;
+import org.apache.flink.connector.jdbc.table.JdbcUpsertTableSink;
+import org.apache.flink.streaming.api.datastream.DataStream;
+import org.apache.flink.streaming.api.datastream.DataStreamSource;
+import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
+import org.apache.flink.table.api.TableSchema;
+import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.operations.ModifyOperation;
+import org.apache.flink.table.operations.Operation;
+import org.apache.flink.table.types.logical.LogicalType;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
 import com.dlink.assertion.Asserts;
 import com.dlink.cdc.AbstractSinkBuilder;
 import com.dlink.cdc.CDCBuilder;
@@ -8,23 +29,6 @@ import com.dlink.executor.CustomTableEnvironment;
 import com.dlink.model.FlinkCDCConfig;
 import com.dlink.model.Schema;
 import com.dlink.model.Table;
-import org.apache.flink.api.common.functions.FilterFunction;
-import org.apache.flink.api.common.serialization.SimpleStringSchema;
-import org.apache.flink.api.dag.Transformation;
-import org.apache.flink.connector.jdbc.JdbcConnectionOptions;
-import org.apache.flink.connector.jdbc.JdbcExecutionOptions;
-import org.apache.flink.connector.jdbc.JdbcSink;
-import org.apache.flink.streaming.api.datastream.DataStreamSource;
-import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.streaming.connectors.kafka.FlinkKafkaProducer;
-import org.apache.flink.table.api.StatementSet;
-import org.apache.flink.table.api.TableException;
-import org.apache.flink.table.operations.ModifyOperation;
-import org.apache.flink.table.operations.Operation;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 /**
  * MysqlCDCBuilder
@@ -54,6 +58,71 @@ public class JdbcSinkBuilder extends AbstractSinkBuilder implements SinkBuilder 
         return new JdbcSinkBuilder(config);
     }
 
+    @Override
+    public void addSink(
+        DataStream<RowData> rowDataDataStream,
+        String schemaTableName,
+        List<String> columnNameList,
+        List<LogicalType> columnTypeList) {
+
+        JdbcUpsertTableSink.Builder builder = JdbcUpsertTableSink.builder();
+
+        Map<String, String> sink = config.getSink();
+        if (sink.containsKey("sink.buffer-flush.interval")) {
+            builder.setFlushIntervalMills(Integer.valueOf(sink.get("sink.buffer-flush.interval")));
+        }
+        if (sink.containsKey("sink.buffer-flush.max-rows")) {
+            builder.setFlushMaxSize(Integer.valueOf(sink.get("sink.buffer-flush.max-rows")));
+        }
+        if (sink.containsKey("sink.max-retries")) {
+            builder.setMaxRetryTimes(Integer.valueOf(sink.get("sink.max-retries")));
+        }
+        JdbcOptions.Builder jdbcOptionsBuilder = JdbcOptions.builder();
+        if (sink.containsKey("connection.max-retry-timeout")) {
+            jdbcOptionsBuilder.setConnectionCheckTimeoutSeconds(Integer.valueOf(sink.get("connection.max-retry-timeout")));
+        }
+        if (sink.containsKey("url")) {
+            jdbcOptionsBuilder.setDBUrl(sink.get("url"));
+        }
+        if (sink.containsKey("dialect")) {
+            switch (sink.get("dialect")) {
+                case "MySql":
+                    jdbcOptionsBuilder.setDialect(new MySQLDialect());
+                    break;
+                case "Oracle":
+                    jdbcOptionsBuilder.setDialect(new OracleDialect());
+                    break;
+                case "ClickHouse":
+                    jdbcOptionsBuilder.setDialect(new ClickHouseDialect());
+                    break;
+                case "SQLServer":
+                    jdbcOptionsBuilder.setDialect(new SQLServerDialect());
+                    break;
+                case "Postgres":
+                    jdbcOptionsBuilder.setDialect(new PostgresDialect());
+                    break;
+            }
+
+        }
+        if (sink.containsKey("driver")) {
+            jdbcOptionsBuilder.setDriverName(sink.get("driver"));
+        }
+        if (sink.containsKey("sink.parallelism")) {
+            jdbcOptionsBuilder.setParallelism(Integer.valueOf(sink.get("sink.parallelism")));
+        }
+        if (sink.containsKey("password")) {
+            jdbcOptionsBuilder.setPassword(sink.get("password"));
+        }
+        if (sink.containsKey("username")) {
+            jdbcOptionsBuilder.setUsername(sink.get("username"));
+        }
+        jdbcOptionsBuilder.setTableName(schemaTableName);
+        builder.setOptions(jdbcOptionsBuilder.build());
+        builder.setTableSchema(TableSchema.fromTypeInfo(rowDataDataStream.getType()));
+        /*JdbcUpsertTableSink build = builder.build();
+        build.consumeDataStream(rowDataDataStream);
+        rowDataDataStream.addSink(build.);*/
+    }
     /*@Override
     public DataStreamSource build(StreamExecutionEnvironment env, DataStreamSource<String> dataStreamSource) {
         final List<Schema> schemaList = config.getSchemaList();
@@ -111,7 +180,7 @@ public class JdbcSinkBuilder extends AbstractSinkBuilder implements SinkBuilder 
             }
             for (Schema schema : schemaList) {
                 for (Table item : schema.getTables()) {
-                    customTableEnvironment.executeSql(item.getFlinkTableSql(sb.toString()+"'table-name' = '"+item.getSchemaTableName()+"'\n"));
+                    customTableEnvironment.executeSql(item.getFlinkTableSql(sb.toString() + "'table-name' = '" + item.getSchemaTableName() + "'\n"));
                     List<Operation> operations = customTableEnvironment.getParser().parse(cdcBuilder.getInsertSQL(item, TABLE_NAME));
                     if (operations.size() > 0) {
                         Operation operation = operations.get(0);
